% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_clm____main.R
\name{cov_clm}
\alias{cov_clm}
\title{Fit covariance structered longitudinal model.}
\usage{
cov_clm(fit, fittingType, sigmaThetaExpr, optimParams)
}
\arguments{
\item{fit}{A fit model with class 'clm'}

\item{fittingType}{A character with the fitting function type. See optins above}

\item{sigmaThetaExpr}{A character with the covariance structure type or a list of expressions}

\item{optimParams}{A list with configuration for optim function. 'Par' is required.}
}
\value{
The fit model with class 'clm.theta'.
}
\description{
Fit covariance structered longitudinal model.
}
\examples{
fitTheta <- cov_clm(fit,
  fittingType = "PML", sigmaThetaExpr = "UCM",
  optimParams = list(par = c(7, 5))
)
}
