% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm__summary.clm____main.R
\name{summary.clm}
\alias{summary.clm}
\title{Summarise the results of 'clm' fit.}
\usage{
\method{summary}{clm}(fit)
}
\arguments{
\item{fit}{A clm fitted model}
}
\value{
Return NULL and print in terminal the results.
}
\description{
Summarise the results of 'clm' fit.
}
\examples{
summary.clm(fit)
}
