% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_clm__summary.clm.theta____main.R
\name{summary.clm.theta}
\alias{summary.clm.theta}
\title{Summarise the results of 'cov_clm' fit.}
\usage{
\method{summary}{clm.theta}(fitTheta)
}
\arguments{
\item{fitTheta}{A clm.theta fitted model}
}
\value{
Return NULL and print in terminal the results.
}
\description{
Summarise the results of 'cov_clm' fit.
}
\examples{
summary.clm.theta(fitTheta)
}
